server_deploy_backup_20260315/backup_deploy/file-agent/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.26.0
openai==1.12.0
pillow==10.2.0
pypdf2==3.0.1
jinja2==3.1.3
python-docx==1.1.0